Fix annoying bug where snake hits itself
The snake hits itself when the user presses the arrow keys too fast,
e.g. if the snake slides down and the user presses the right and up
arrow, only the up event is processed and the snake slides upwards
into itself. In such a case the update is ignored and the snake will
continue to slide downwards.